Refactor tooltip positioning logic into a separate utility class and update WnTooltip to utilize it. Modify button navigation in NetworkScreen for improved routing.
